Added SignIn, SignUp and Forgot Password Activity and created account with Firebase. Also added Password Visibilty Toggle.
