+++ app/src/main/AndroidManifest.xml
@@ -17,6 +17,9 @@
             </intent-filter>
         </activity>
         <activity android:name=".MainActivity"></activity>
+        <activity android:name=".accounts.SignupActivity" />
+        <activity android:name=".accounts.SigninActivity" />
+        <activity android:name=".accounts.ForgotPasswordActivity" />
     </application>
 
 </manifest>
