--- app/src/main/res/layout/activity_sign_up.xml
@@ -1,141 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".SignUpActivity"
-    android:orientation="vertical"
-    android:background="@color/colorPrimary"
-    android:id="@+id/linearlayout">
-
-    <ImageView
-        android:layout_marginTop="32dp"
-        android:layout_width="@dimen/ImageWidth"
-        android:layout_height="@dimen/ImageHeight"
-        android:layout_gravity="center"
-        android:layout_marginBottom="@dimen/MarginBottom"
-        android:src="@drawable/woman_power_emblem"
-        android:visibility="visible"/>
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/welcome_to_nirbheek"
-        android:textAllCaps="true"
-        android:layout_gravity="center_horizontal"
-        android:textColor="@color/TextColor"
-        android:layout_marginBottom="@dimen/TextMarginBottom"
-        android:textSize="@dimen/TextSize"
-        android:textStyle="bold"/>
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Sign Up to Continue"
-        android:textAllCaps="true"
-        android:layout_gravity="center_horizontal"
-        android:textColor="@color/textColor"
-        android:layout_marginBottom="@dimen/MarginBottom"/>
-
-    <ProgressBar
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:id="@+id/probar"
-        android:layout_gravity="center_horizontal"
-        android:visibility="gone"
-        />
-
-    <LinearLayout
-        android:id="@+id/jadu_layout"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="vertical"
-        android:layout_marginBottom="@dimen/MarginBottom"
-        >
-
-
-        <EditText
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:hint="Full Name "
-            android:id="@+id/edName"
-            android:inputType="textEmailAddress"
-            android:fontFamily="sans-serif"
-            android:textSize="@dimen/EmailTextSize"
-            android:layout_marginRight="@dimen/MarginRight"
-            android:layout_marginLeft="@dimen/MarginLeft"
-            android:layout_marginTop="@dimen/MarginTop"
-
-            />
-
-        <EditText
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:hint="Enter Email Id "
-            android:id="@+id/edEmail"
-            android:inputType="textEmailAddress"
-            android:fontFamily="sans-serif"
-            android:textSize="@dimen/EmailTextSize"
-            android:layout_marginRight="@dimen/MarginRight"
-            android:layout_marginLeft="@dimen/MarginLeft"
-            />
-        <EditText
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:hint="Phone Number "
-            android:id="@+id/edPhone"
-            android:inputType="textEmailAddress"
-            android:fontFamily="sans-serif"
-            android:textSize="@dimen/EmailTextSize"
-            android:layout_marginRight="@dimen/MarginRight"
-            android:layout_marginLeft="@dimen/MarginLeft"
-            />
-
-        <android.support.design.widget.TextInputLayout
-            android:id="@+id/text_input_layout"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            app:passwordToggleEnabled="true"
-            >
-
-        <android.support.design.widget.TextInputEditText
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginRight="@dimen/MarginRight"
-            android:layout_marginLeft="@dimen/MarginLeft"
-            android:inputType="textPassword"
-            android:id="@+id/edPass"
-            android:hint="Password"
-            />
-        </android.support.design.widget.TextInputLayout>
-
-        <Button
-            android:id="@+id/BtnSignUp"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center"
-            android:text="@string/sign_up"
-            android:layout_marginBottom="@dimen/MarginBottom"
-            android:layout_marginTop="@dimen/MarginTop"
-            android:layout_marginRight="@dimen/MarginRight"
-            android:layout_marginLeft="@dimen/MarginLeft"
-            android:background="@color/ButtonBackground"
-            />
-
-        <TextView
-            android:id="@+id/TxtLogin1"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center"
-            android:text="Already Have An Account?Login"
-            android:textSize="@dimen/TextSize"
-            android:textColor="@color/textColor"
-            />
-
-
-
-    </LinearLayout>
-
-
-</LinearLayout>
